colocate project section in profile dir
